Refactor to separate concerns:
 - separated ScalaJS/canvas renderer into its own package;
 - separated the code for BoundingBox from Image.
