+++ project/build.sbt
@@ -1,4 +1,4 @@
-addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.0-M3")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.0-M1")
 
 addSbtPlugin("com.lihaoyi" % "workbench" % "0.2.3")
