+++ src/main/scala/doodle/canvas/Draw.scala
@@ -1,11 +1,17 @@
-package doodle
+package doodle.canvas
 
 import org.scalajs.dom
 
+import doodle.core._
+
 object Draw {
+  def apply(img: Image, id: String): Unit = {
+    apply(img, dom.document.getElementById("canvas").asInstanceOf[dom.HTMLCanvasElement])
+  }
+
   def apply(img: Image, canvas: dom.HTMLCanvasElement): Unit = {
     val ctx = canvas.getContext("2d").asInstanceOf[dom.CanvasRenderingContext2D]
-    val originX  = canvas.width / 2
+    val originX = canvas.width / 2
     val originY = canvas.height / 2
 
     draw(img, originX, originY, DrawingContext.blackLines, ctx)
@@ -53,9 +59,9 @@ object Draw {
         draw(b, originX, originY, context, ctx)
         draw(t, originX, originY, context, ctx)
       case b @ Beside(l, r) =>
-        val box = b.boundingBox
-        val lBox = l.boundingBox
-        val rBox = r.boundingBox
+        val box = BoundingBox(b)
+        val lBox = BoundingBox(l)
+        val rBox = BoundingBox(r)
 
         val lOriginX = originX + box.left + (lBox.width / 2)
         val rOriginX = originX + box.right - (rBox.width / 2)
@@ -65,9 +71,9 @@ object Draw {
         draw(l, lOriginX, originY, context, ctx)
         draw(r, rOriginX, originY, context, ctx)
       case a @ Above(t, b) =>
-        val box = a.boundingBox
-        val tBox = t.boundingBox
-        val bBox = b.boundingBox
+        val box = BoundingBox(a)
+        val tBox = BoundingBox(t)
+        val bBox = BoundingBox(b)
 
         val tOriginY = originY + box.top + (tBox.height / 2)
         val bOriginY = originY + box.bottom - (bBox.height / 2)
